requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
django-ckeditor==6.7.0
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sorl-thumbnail==12.10.0
gunicorn==21.2.0
whitenoise==6.6.0